update-entity!: ignore virtual fields
Before this change, because of `#:when (meta-can-update? (entity-meta
entity))` in `update!`, the `#f` condition in `update-entity!` seems
to have been unreachable, and also it would violate the contract on
`update!`, which only allows entities in a list; but with the addition
of virtual fields, the false condition became useful again - for cases
when an entity has not been updated despite having been
modified (because the modificationts are on virtual fields only).

So I have moved the check inside `update-entity!`:

```diff
   (cond
-    [(entity-changed? entity)
+    [(meta-can-update? (entity-meta entity))
```

It seems to me that `entity-changed?` doubles `meta-can-update?`, so I
have only left the latter.